[Bronze IV] Title: Quality-Adjusted Life-Year, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/22279. Quality－Adjusted Life－Year/Quality－Adjusted Life－Year.cc
@@ -1,13 +1,13 @@
 #include <cstdio>
 
 int N;
-double q, y, s;
+float q, y, s;
 
 int main() {
   scanf("%d", &N);
   while (N--) {
-    scanf("%lf%lf", &q, &y);
+    scanf("%f%f", &q, &y);
     s += q * y;
   }
-  printf("%.3lf", s);
+  printf("%.3f", s);
 }
